+++ Data/Data.Models/Models/EmployeeDepartmentHistory.cs
@@ -20,7 +20,7 @@ namespace Data.Models
         /// <summary>
         /// Identifies which 8-hour shift the employee works. Foreign key to Shift.Shift.ID.
         /// </summary>
-        public byte ShiftId { get; set; }
+        public int ShiftId { get; set; }
         /// <summary>
         /// Date the employee started work in the department.
         /// </summary>
